+++ tests/compas_timber/test_model.py
@@ -15,10 +15,10 @@ def test_create():
     assert model
 
 
-def test_add_beam():
+def test_add_element():
     A = TimberModel()
     B = Beam(Frame.worldXY(), width=0.1, height=0.1, length=1.0)
-    A.add_beam(B)
+    A.add_element(B)
 
     assert B in A.beams
     assert len(list(A.graph.nodes())) == 1
@@ -32,8 +32,8 @@ def test_add_joint():
     b1 = Beam(Frame.worldXY(), length=1.0, width=0.1, height=0.1)
     b2 = Beam(Frame.worldYZ(), length=1.0, width=0.1, height=0.1)
 
-    model.add_beam(b1)
-    model.add_beam(b2)
+    model.add_element(b1)
+    model.add_element(b2)
     _ = LButtJoint.create(model, b1, b2)
 
     assert len(model.beams) == 2
@@ -47,8 +47,8 @@ def test_copy(mocker):
     B1 = Beam(F1, length=1.0, width=0.1, height=0.12)
     B2 = Beam(F2, length=1.0, width=0.1, height=0.12)
     A = TimberModel()
-    A.add_beam(B1)
-    A.add_beam(B2)
+    A.add_element(B1)
+    A.add_element(B2)
     _ = LButtJoint.create(A, B1, B2)
 
     A_copy = A.copy()
@@ -63,8 +63,8 @@ def test_deepcopy(mocker):
     B1 = Beam(F1, length=1.0, width=0.1, height=0.12)
     B2 = Beam(F2, length=1.0, width=0.1, height=0.12)
     A = TimberModel()
-    A.add_beam(B1)
-    A.add_beam(B2)
+    A.add_element(B1)
+    A.add_element(B2)
     _ = LButtJoint.create(A, B1, B2)
 
     A_copy = A.copy()
@@ -77,9 +77,9 @@ def test_beams_have_keys_after_serialization():
     B1 = Beam(Frame.worldXY(), length=1.0, width=0.1, height=0.1)
     B2 = Beam(Frame.worldYZ(), length=1.0, width=0.1, height=0.1)
     B3 = Beam(Frame.worldZX(), length=1.0, width=0.1, height=0.1)
-    A.add_beam(B1)
-    A.add_beam(B2)
-    A.add_beam(B3)
+    A.add_element(B1)
+    A.add_element(B2)
+    A.add_element(B3)
     keys = [beam.guid for beam in A.beams]
 
     A = json_loads(json_dumps(A))
@@ -94,8 +94,8 @@ def test_serialization_with_l_butt_joints(mocker):
     B1 = Beam(F1, length=1.0, width=0.1, height=0.12)
     B2 = Beam(F2, length=1.0, width=0.1, height=0.12)
     A = TimberModel()
-    A.add_beam(B1)
-    A.add_beam(B2)
+    A.add_element(B1)
+    A.add_element(B2)
     _ = LButtJoint.create(A, B1, B2)
 
     A = json_loads(json_dumps(A))
@@ -106,8 +106,8 @@ def test_serialization_with_t_butt_joints(mocker):
     a = TimberModel()
     b1 = Beam(Frame.worldXY(), length=1.0, width=0.1, height=0.1)
     b2 = Beam(Frame.worldYZ(), length=1.0, width=0.1, height=0.1)
-    a.add_beam(b1)
-    a.add_beam(b2)
+    a.add_element(b1)
+    a.add_element(b2)
     _ = TButtJoint.create(a, b1, b2)
 
     a = json_loads(json_dumps(a))