+++ CHANGELOG.md
@@ -9,9 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `Plate` element.
+* Added attribute `plates` to `TimberModel`.
+
 ### Changed
 
+* Renamed `beam` to `element` in different locations to make it more generic.
+
 ### Removed
+
+* Removed `add_beam` from `TimberModel`, use `add_element` instead.
+* Removed `add_plate` from `TimberModel`, use `add_element` instead.
+* Removed `add_wall` from `TimberModel`, use `add_element` instead.
 
 
 ## [0.9.1] 2024-07-05